Address comments, remove currently-unused utils for model versions
